refactor(paths): consolidate coven-home and engine-root resolution
Four independent implementations of "where is coven home?" had drifted:
main.rs owned the robust chain (COVEN_HOME -> HOME -> USERPROFILE ->
HOMEDRIVE/HOMEPATH -> platform home), while the chat TUI client silently
fell back to a cwd-relative .coven, harness.rs skipped the Windows
profile chain, and engine.rs/engine_install.rs each hand-built
~/.coven/engine. Chat exports hardcoded ~/.coven/exports, ignoring
COVEN_HOME.

New paths module owns the policy: coven_home_dir() (moved from main.rs,
re-exported so call sites are unchanged) and managed_engine_root(),
which stays keyed on the user home by design (per-user engine install,
documented in the module docs). Callers now delegate:

- tui/chat/client.rs: drifted local resolver and Default impl removed;
  DaemonChatClient::detect() fails closed instead of guessing a
  cwd-relative home. App::new is fallible; shell.rs cast paths use
  detect()?.
- tui/chat/app.rs: observe/doctor/export views resolve through one
  resolved_coven_home() helper; exports now land under the real Coven
  home instead of a hardcoded ~/.coven.
- harness.rs coven_home_from_process_env delegates (gains the full
  Windows fallback chain).
- engine.rs / engine_install.rs share managed_engine_root().

Tests move to paths.rs and add coverage for empty-env filtering and the
no-home failure.

Closes #396
